pyproject.toml
[tool.poetry]
name = "foods"
version = "0.1.0"
description = ""
authors = ["Aratimius <aratimius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
ipython = "^8.32.0"
black = "^25.1.0"
python-dotenv = "^1.0.1"
django-model-utils = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
